+++ Libraries/Network/Packets/Structures/NewEntity.hpp
@@ -15,9 +15,9 @@
  * 
  */
 struct packet_new_entity {
-    uint16_t id; // id of new entity;
-    uint16_t positionX; // x position of new entity;
-    uint16_t positionY; // y position of new entity;
+    int id; // id of new entity;
+    float positionX; // x position of new entity;
+    float positionY; // y position of new entity;
     uint16_t direction; // direction of new entity;
     uint16_t type; // type of entity;
 };
